+++ examples/data-solutions/gcs-to-bq-with-dataflow/variables.tf
@@ -16,6 +16,7 @@
 variable "prefix" {
   description = "Unique prefix used for resource names. Not used for project if 'project_create' is null."
   type        = string
+  default     = null
 }
 
 variable "project_create" {